requirements.txt
Flask
loguru
openai
python-dotenv